+++ packages/manager/apps/pci-ai-tools/src/__tests__/helpers/mocks/app/appHelper.ts
@@ -1,0 +1,16 @@
+import ai from '@/types/AI';
+
+export const mockedFixedScaling: ai.app.ScalingStrategy = {
+  fixed: {
+    replicas: 2,
+  },
+};
+
+export const mockedAutoScaling: ai.app.ScalingStrategy = {
+  automatic: {
+    replicasMin: 2,
+    replicasMax: 100,
+    averageUsageTarget: 75,
+    resourceType: ai.app.ScalingAutomaticStrategyResourceTypeEnum.CPU,
+  },
+};