+++ packages/manager/apps/pci-ai-tools/src/pages/dashboard/docker/_components/privateDocker/_components/docker.constant.ts
@@ -1,0 +1,12 @@
+export const DOCKER_CONFIG = {
+  region: {
+    min: 3,
+    max: 3,
+  },
+  other: {
+    min: 1,
+  },
+  name: {
+    pattern: /^[\w_][\w\-_]{0,60}$/,
+  },
+};