+++ packages/manager/apps/freefax/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.0.24](https://github.com/ovh/manager/compare/@ovh-ux/manager-freefax-app@9.0.23...@ovh-ux/manager-freefax-app@9.0.24) (2025-04-28)
+
+**Note:** Version bump only for package @ovh-ux/manager-freefax-app
+
+
+
+
+
 ## [9.0.23](https://github.com/ovh/manager/compare/@ovh-ux/manager-freefax-app@9.0.22...@ovh-ux/manager-freefax-app@9.0.23) (2025-04-22)
 
 **Note:** Version bump only for package @ovh-ux/manager-freefax-app